Migrated code of homeFragment to TaskFragment
