+++ app/src/main/java/com/timeit/app/Fragments/TaskFragment.kt
@@ -1,60 +1,124 @@
 package com.timeit.app.Fragments
 
+import android.app.DatePickerDialog
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.GridLayoutManager
+import com.timeit.app.Adapters.DateAdapter
+import com.timeit.app.DataModels.Day
 import com.timeit.app.R
+import com.timeit.app.Utils
+import com.timeit.app.databinding.FragmentTaskBinding
+import java.text.SimpleDateFormat
+import java.util.Calendar
+import java.util.Locale
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [TaskFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class TaskFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
-    }
-
+    private var _binding: FragmentTaskBinding? = null
+    private val binding get() = _binding!!
+    private val utils = Utils()
+    private lateinit var dateAdapter: DateAdapter
+    private var selectedDayPosition = -1
+    private var currentWeekStart: Calendar = Calendar.getInstance()
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_task, container, false)
+        _binding = FragmentTaskBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        dateAdapter = DateAdapter(generateWeekDays(currentWeekStart))
+        binding.recyclerView.layoutManager = GridLayoutManager(context, 7)
+        binding.recyclerView.adapter = dateAdapter
+
+        binding.leftArrow.setOnClickListener { showPreviousWeek() }
+        binding.rightArrow.setOnClickListener { showNextWeek() }
+
+        val today = Calendar.getInstance().get(Calendar.DAY_OF_WEEK) - 1
+        dateAdapter.updateSelected(today)
+
+        val todayPosition = getTodayPosition(dateAdapter.dateItemList)
+        selectedDayPosition = todayPosition
+        dateAdapter.updateSelected(todayPosition)
+        binding.recyclerView.scrollToPosition(todayPosition)
+
+        dateAdapter.setOnItemClickListener { position ->
+            selectedDayPosition = position
+            dateAdapter.updateSelected(position)
+            val selectedDate = dateAdapter.dateItemList[position]
+            binding.selectedDayText.text = "${selectedDate.dayMonth} ${selectedDate.year}"
+        }
+
+        binding.monthtext.setOnClickListener {
+            showDatePickerDialog()
+        }
+
+        setCurrentMonth(currentWeekStart)
+    }
+
+    private fun generateWeekDays(startingDay: Calendar): List<Day> {
+        return utils.generateDaysForWeek(startingDay)
+    }
+
+    private fun showPreviousWeek() {
+        currentWeekStart.add(Calendar.WEEK_OF_YEAR, -1)
+        updateWeek()
+    }
+
+    private fun showNextWeek() {
+        currentWeekStart.add(Calendar.WEEK_OF_YEAR, 1)
+        updateWeek()
+    }
+
+    private fun updateWeek() {
+        dateAdapter.updateData(generateWeekDays(currentWeekStart))
+        binding.recyclerView.scrollToPosition(0)
+        setCurrentMonth(currentWeekStart)
+    }
+
+    private fun getTodayPosition(dates: List<Day>): Int {
+        return dates.indexOfFirst { it.isToday }
+    }
+
+    private fun setCurrentMonth(calendar: Calendar) {
+        val monthYearFormat = SimpleDateFormat("MMMM yyyy", Locale.getDefault())
+        val monthYearText = monthYearFormat.format(calendar.time)
+        binding.selectedDayText.text = monthYearText
+    }
+
+    private fun showDatePickerDialog() {
+        val calendar = Calendar.getInstance()
+        val year = calendar.get(Calendar.YEAR)
+        val month = calendar.get(Calendar.MONTH)
+        val day = calendar.get(Calendar.DAY_OF_MONTH)
+
+        val datePickerDialog = DatePickerDialog(requireContext(), { _, selectedYear, selectedMonth, selectedDay ->
+            val selectedCalendar = Calendar.getInstance()
+            selectedCalendar.set(selectedYear, selectedMonth, selectedDay)
+            currentWeekStart = selectedCalendar
+            updateWeek()
+            val dateText = SimpleDateFormat("MMMM yyyy", Locale.getDefault()).format(selectedCalendar.time)
+            binding.selectedDayText.text = dateText
+        }, year, month, day)
+
+        datePickerDialog.show()
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
     }
 
     companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment TaskFragment.
-         */
-        // TODO: Rename and change types and number of parameters
         @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            TaskFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
-            }
+        fun newInstance() = TaskFragment()
     }
 }
