+++ app/src/main/res/layout/fragment_habits.xml
@@ -4,11 +4,15 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
+    android:gravity="center"
     tools:context=".Fragments.HabitsFragment">
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerViewHabits"
+    <TextView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
+        android:layout_height="wrap_content"
+        android:text="habits fragment"
+        android:textStyle="bold"
+        android:textSize="30dp"
+        android:gravity="center" />
 
 </LinearLayout>
