+++ app/src/main/res/layout/fragment_home.xml
@@ -34,9 +34,10 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="20dp">
+        android:layout_margin="10dp">
 
         <com.google.android.material.tabs.TabLayout
+            android:id="@+id/tabMode"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             app:tabGravity="fill"
@@ -61,90 +62,9 @@
         </com.google.android.material.tabs.TabLayout>
     </LinearLayout>
 
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_marginTop="20dp"
-        >
-
-        <LinearLayout
-            android:id="@+id/monthtext"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:gravity="center_vertical">
-
-            <TextView
-                android:id="@+id/selectedDayText"
-                style="@style/TitleSelectedDate"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentStart="true"
-                android:layout_centerHorizontal="true"
-                android:layout_centerVertical="true"
-                android:layout_marginStart="20dp"
-                android:fontFamily="@font/plusjakarta_semibold"
-                android:paddingTop="8dp"
-                android:paddingBottom="8dp"
-                android:textSize="20sp"
-                android:textStyle="bold"
-                tools:text="August, 2024" />
-            <ImageView
-                android:id="@+id/monthArrow"
-                android:layout_width="10dp"
-                android:layout_height="10dp"
-                android:rotation="90"
-                android:layout_marginStart="10dp"
-                android:layout_gravity="center_vertical"
-                android:src="@drawable/ic_weekday_chevron_down"
-                />
-        </LinearLayout>
-
-
-        <LinearLayout
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentEnd="true"
-            android:layout_centerInParent="true"
-            android:layout_marginRight="20dp"
-            >
-            <ImageView
-                android:id="@+id/leftArrow"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:src="@drawable/ic_weekday_chevron_left"
-                android:padding="16dp"
-             />
-            <ImageView
-                android:id="@+id/rightArrow"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:src="@drawable/ic_weekday_chevron_right"
-                android:padding="16dp"
-                />
-
-        </LinearLayout>
-    </RelativeLayout>
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView"
-        android:paddingStart="12dp"
-        android:paddingEnd="12dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        android:orientation="horizontal"
-        android:scrollbars="none"
-        tools:listitem="@layout/item_date"/>
-
-
     <FrameLayout
         android:id="@+id/fragmentContainer"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_weight="0.4"
-        android:layout_marginTop="20dp"/>
+        android:layout_height="match_parent" />
 
 </LinearLayout>
