+++ app/src/main/res/layout/item_date.xml
@@ -4,8 +4,8 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:layout_marginRight="10dp"
-    android:padding="10dp"
+    android:layout_marginRight="5dp"
+    android:padding="8dp"
     android:background="@android:color/transparent">
 
         <TextView
